Moved data loading methods into own module, outside of test_support.
