+++ fc/test_support.py
@@ -6,7 +6,7 @@ Routines of use in tests of Functional Curation.
 import os
 import numpy as np
 
-from .language import values as V
+from .data_loading import load, load2d
 
 
 def get_process_number():
@@ -150,39 +150,4 @@ def check_results(proto, expected_spec, data_folder, rel_tol=0.01, abs_tol=0, me
                          bad_entries.nonzero()[:10]))
                     results_ok = False
     return results_ok
-
-
-def check_file_compression(file_path):
-    """Return (real_path, is_compressed) if a .gz compressed version of file_path exists."""
-    real_path = file_path
-    is_compressed = False
-    if file_path.endswith('.gz'):
-        is_compressed = True
-    elif os.path.exists(file_path + '.gz'):
-        real_path += '.gz'
-        is_compressed = True
-    return real_path, is_compressed
-
-
-def load2d(file_path):
-    """Load the legacy data format for 2d arrays."""
-    real_path, is_compressed = check_file_compression(file_path)
-    array = np.loadtxt(real_path, dtype=float, delimiter=',', ndmin=2, unpack=True)  # unpack transposes the array
-    assert array.ndim == 2
-    return V.Array(array)
-
-
-def load(file_path):
-    """Load the legacy data format for arbitrary dimension arrays."""
-    real_path, is_compressed = check_file_compression(file_path)
-    if is_compressed:
-        import gzip
-        f = gzip.GzipFile(real_path, 'rb')
-    else:
-        f = open(real_path, 'r')
-    f.readline()  # strip comment line
-    dims = map(int, f.readline().split(','))[1:]
-    array = np.loadtxt(f, dtype=float)
-    f.close()
-    return V.Array(array.reshape(tuple(dims)))
